Rename cli to cmd
